moved redux to the correct files!
