+++ client/components/Account.js
@@ -4,13 +4,13 @@ import {connect} from 'react-redux'
 /**
  * COMPONENT
  */
-export const Home = props => {
+export const Account = props => {
   const {username} = props
 
   return (
     <div>
       <h3>Welcome back {username}! </h3>
-    
+
     </div>
   )
 }
@@ -24,4 +24,4 @@ const mapState = state => {
   }
 }
 
-export default connect(mapState)(Home)
+export default connect(mapState)(Account)
