+++ client/components/Navbar.js
@@ -10,8 +10,8 @@ const Navbar = ({ handleClick, isLoggedIn }) => (
       {isLoggedIn ? (
         <div className='navbar'>
           {/* The navbar will show these links after you log in */}
-          <Link to='/homepage'>Home</Link>
-          <Link to='/home'>Home</Link>
+          <Link to='/'>Home</Link>
+          <Link to='/account'>Account</Link>
           <Link to='/rooms'>Rooms</Link>
           <Link to='/experiences'>Experiences</Link>
           <a href='#' onClick={handleClick}>
@@ -23,11 +23,13 @@ const Navbar = ({ handleClick, isLoggedIn }) => (
       ) : (
         <div className='navbar'>
           {/* The navbar will show these links before you log in */}
-          <Link to='/homepage'>Home</Link>
+          <Link to='/'>Home</Link>
+          <Link to='/account'>Account</Link>
           <Link to='/rooms'>Rooms</Link>
           <Link to='/experiences'>Experiences</Link>
           <Link to='/login'>Login</Link>
           <Link to='/signup'>Sign Up</Link>
+          <div className='button'> CHECK OUT <i className='fa fa-shopping-cart'></i></div>
         </div>
       )}
     </nav>
