+++ client/store/singleExperience.js
@@ -1,17 +1,12 @@
 import axios from 'axios';
 
 const GET_SINGLE_EXPERIENCE = 'GET_SINGLE_EXPERIENCE';
-const UPDATE_EXPERIENCE = 'UPDATE_EXPERIENCE';
 
 const getSingleExperience = (experience) => {
   return {
     type: GET_SINGLE_EXPERIENCE,
     experience,
   };
-};
-
-const _updateExperience = (newExperience) => {
-  return { type: UPDATE_EXPERIENCE, updatedExperience: newExperience };
 };
 
 export const fetchSingleExperience = (experienceId) => {
@@ -21,33 +16,12 @@ export const fetchSingleExperience = (experienceId) => {
   };
 };
 
-export const updateExperience = (id, experienceToUpdate, history) => {
-  return async (dispatch) => {
-    const response = await axios.put(
-      `/api/experiences/${id}`,
-      experienceToUpdate
-    );
-    const updatedExperience = response.data;
-    dispatch(_updateExperience(updatedExperience));
-    history.push('/');
-  };
-};
-
 const initialState = {};
 
 const singleExperienceReducer = (state = initialState, action) => {
   switch (action.type) {
     case GET_SINGLE_EXPERIENCE:
       return action.experience;
-    case UPDATE_EXPERIENCE:
-      let newExperiences = [...state];
-      newExperiences = newExperiences.map((experience) => {
-        if (experience.id === action.updatedExperience.id) {
-          return action.updatedExperience;
-        }
-        return experience;
-      });
-      return { ...state, newExperiences };
     default:
       return state;
   }
