+++ client/store/singleRoom.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 ////ACTION TYPES
 const SET_ROOM = 'SET_ROOM';
 const TOGGLE_AVAILABLE = 'TOGGLE_AVAILABLE';
-const UPDATE_ROOM = 'UPDATE_EXPERIENCE';
 
 //ACTION CREATORS
 const setRoom = (room) => {
@@ -17,10 +16,6 @@ const toggleAvailable = (availability) => {
     type: TOGGLE_AVAILABLE,
     availability,
   };
-};
-
-const _updateRoom = (newRoom) => {
-  return { type: UPDATE_ROOM, updatedRoom: newRoom };
 };
 
 //THUNK CREATORS
@@ -48,30 +43,12 @@ export const toggleStatus = (id, availability) => {
   };
 };
 
-export const updateRoom = (id, roomToUpdate, history) => {
-  return async (dispatch) => {
-    const response = await axios.put(`/api/experiences/${id}`, roomToUpdate);
-    const updatedRoom = response.data;
-    dispatch(_updateRoom(updatedRoom));
-    history.push('/');
-  };
-};
-
 export default function singleRoomReducer(state = {}, action) {
   switch (action.type) {
     case SET_ROOM:
       return action.room;
     case TOGGLE_AVAILABLE:
       return { ...state, available: action.availability };
-    case UPDATE_ROOM:
-      let newRooms = [...state];
-      newRooms = newRooms.map((room) => {
-        if (room.id === action.updatedRoom.id) {
-          return action.updatedRoom;
-        }
-        return room;
-      });
-      return { ...state, newRooms };
     default:
       return state;
   }
